% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_env_mix.R
\name{gen_env_mix}
\alias{gen_env_mix}
\title{Title}
\usage{
gen_env_mix(
  n,
  p,
  q,
  error.dist = "mvnorm",
  alpha,
  sigma,
  return.T = TRUE,
  mu.X = NULL,
  return.full.data = FALSE,
  LOD = NULL,
  LOD.p = NULL
)
}
\arguments{
\item{sigma}{}
}
\value{

}
\description{
Title
}
