% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LODmix-package.R
\docType{package}
\name{LODmix-package}
\alias{LODmix}
\alias{LODmix-package}
\title{LODmix: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vsritter/LODmix}
  \item Report bugs at \url{https://github.com/vsritter/LODmix/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
